+++ src/h3/router.ts
@@ -1,7 +1,10 @@
-import { ControllerAction, HttpMethod, RouteInfo } from "../../types/basic";
-import { H3App, Handler, HttpContext, Middleware } from "../../types/h3";
+import { ClearRequest } from "src/ClearRequest";
+import { ControllerAction, HttpMethod, RouteInfo } from "types/basic";
+import { H3App, Handler, HttpContext, Middleware, RouteHandler } from "types/h3";
 
-import { type H3 } from "h3";
+import { getQuery, getRouterParams, readBody, type H3 } from "h3";
+import { Controller } from "src/Controller";
+import { Route } from "src/Route";
 
 /**
  * @class clear-router
@@ -13,12 +16,7 @@ export class Router {
     /**
      * All registered routes
      */
-    static routes: Array<{
-        methods: HttpMethod[];
-        path: string;
-        handler: Handler;
-        middlewares: Middleware[];
-    }> = [];
+    static routes: Array<Route<HttpContext, Middleware>> = [];
 
     /**
      * Current route prefix
@@ -63,16 +61,12 @@ export class Router {
         const methodArray = Array.isArray(methods) ? methods : [methods]
         const fullPath = this.normalizePath(`${this.prefix}/${path}`)
 
-        this.routes.push({
-            methods: methodArray,
-            path: fullPath,
+        this.routes.push(new Route(
+            methodArray,
+            fullPath,
             handler,
-            middlewares: [
-                ...this.globalMiddlewares,
-                ...this.groupMiddlewares,
-                ...(middlewares || []),
-            ],
-        })
+            [...this.globalMiddlewares, ...this.groupMiddlewares, ...(middlewares || [])]
+        ))
     }
 
     /**
@@ -239,11 +233,15 @@ export class Router {
      */
     static apply (app: H3): H3App {
         for (const route of this.routes) {
-            let handlerFunction: ((ctx: HttpContext) => any | Promise<any>) | null = null
+            let handlerFunction: RouteHandler | null = null
+            let instance: Controller<HttpContext> | null = null
 
             try {
                 if (typeof route.handler === 'function') {
-                    handlerFunction = route.handler
+                    /** 
+                     * Since we do not have a controller instance, we will call the handler function directly and the route instance will be the this argument. This allows for both controller-based and function-based handlers to work seamlessly.
+                     */
+                    handlerFunction = route.handler.bind(route)
                 } else if (
                     Array.isArray(route.handler) &&
                     route.handler.length === 2
@@ -251,15 +249,16 @@ export class Router {
                     const [Controller, method] = route.handler
 
                     if (
-                        typeof Controller === 'function' &&
+                        ['function', 'object'].includes(typeof Controller) &&
                         typeof Controller[method] === 'function'
                     ) {
+                        instance = Controller
                         handlerFunction = Controller[method].bind(Controller)
                     }
                     else if (typeof Controller === 'function') {
-                        const instance = new Controller()
-                        if (typeof instance[method] === 'function') {
-                            handlerFunction = instance[method].bind(instance)
+                        instance = new Controller()
+                        if (typeof instance![method] === 'function') {
+                            handlerFunction = instance![method].bind(instance)
                         } else {
                             throw new Error(
                                 `Method "${method}" not found in controller instance "${Controller.name}"`
@@ -300,7 +299,8 @@ export class Router {
                 app[method](route.path, async (event) => {
                     try {
                         const ctx: HttpContext = event
-                        const result = handlerFunction(ctx)
+                        await Router.bindRequestToInstance(ctx, instance ?? route)
+                        const result = handlerFunction(ctx, instance?.clearRequest!)
                         return await Promise.resolve(result)
                     } catch (error: any) {
                         return error
@@ -312,6 +312,24 @@ export class Router {
         }
 
         return app as never
+    }
+
+    private static async bindRequestToInstance (
+        ctx: HttpContext,
+        instance: Controller<HttpContext> | Route<HttpContext, Middleware> | null
+    ): Promise<void> {
+        if (!instance) return
+
+        instance.ctx = ctx
+        instance.body = await readBody(ctx) ?? {}
+        instance.query = getQuery(ctx)
+        instance.params = getRouterParams(ctx, { decode: true })
+        instance.clearRequest = new ClearRequest({
+            ctx,
+            body: instance.body,
+            query: instance.query,
+            params: instance.params,
+        })
     }
 }