+++ types/h3.ts
@@ -1,5 +1,6 @@
 import type { H3, H3Event, TypedServerRequest } from 'h3';
 
+import { ClearRequest } from 'src/ClearRequest';
 import type { ControllerHandler } from './basic';
 
 export type H3App = Omit<H3['fetch'], 'fetch'> & { fetch: (request: TypedServerRequest) => Promise<Response> }
@@ -15,7 +16,14 @@ export interface HttpContext extends H3Event { }
  * Route handler function type
  */
 export type RouteHandler = (
+    /**
+     * H3 event context
+     */
     ctx: HttpContext,
+    /**
+     * ClearRequest instance
+     */
+    req: ClearRequest
 ) => any | Promise<any>;
 
 /**